deploy/backup_auth/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
SQLAlchemy>=2.0,<3.0
PyMySQL>=1.1,<2.0
passlib[bcrypt]>=1.7,<2.0
bcrypt==4.0.1
PyJWT>=2.8,<3.0
